% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs_quan.R
\name{exposure.prior}
\alias{exposure.prior}
\title{Construct design matrices for estimated exposure quantile functions}
\usage{
exposure.prior(inde, re, burn_in, ...)
}
\arguments{
\item{inde}{a logical value indicating whether exposure quantile processes are 
assumed to be independent across time points.}

\item{re}{the output from applying the function \code{\link{fit.exposure}}.}

\item{burn_in}{a number specifying the number of iterations that is discarded.}
}
\value{
This function returns a list containing: 
(1) a list include vectors containing the mean of the multivariate normal (MVN) 
prior assumed for basis coefficients used for modeling exposure quantile functions (\code{theta.pri.mu}), 
(2) a list include matrices containing the precision matrix of the MVN prior (\code{theta.pri.pre}).
}
\description{
This function constructs design matrices corresponding to estimated 
exposure quantile functions, which will be used subsequently for 
fitting negative binomial regression models using the function \code{fit.health.quan.errors}.
}
\examples{
# Simulate time-specific means and SDs of exposure quantile functions
mean.vec <- rnorm(10, mean = 7.2, sd = 1)
sd.vec <- rnorm(10, mean = 1, sd = 0.2)

# Simulate individual-level exposures for all 10 time points
x.sim.mat <- do.call(rbind, lapply(1:10, 
function(x) rnorm(num.size.time, mean.vec[x], sd.vec[x])))

# Estimate exposure quantile functions assuming exposure quantile functions 
# are independent across time points.
re.fit.exp <- NULL
for(i in 1:10){
re.fit.exp[[i]] <- fit.exposure(x.ind = x.sim.mat[i, ],
basis.fun = "Gau", L = 4, niter = 10000, inde = T) }

# Obtain means and precision matrices of the MVN prior.
MVNprior.mu.pre.ind <- exposure.prior(re = re.fit.exp, burn_in = 5000, inde = T)

}
