% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs_quan.R
\name{fit.exposure}
\alias{fit.exposure}
\title{Estimate exposure quantile functions using semiparametric Bayesian approaches}
\usage{
fit.exposure(x.ind, basis.fun, basis.shape = 5, L = 4, niter, inde = TRUE, ...)
}
\arguments{
\item{x.ind}{a matrix with each row contains individual-level exposures.}

\item{basis.fun}{a character string specifying which basis functions are used 
for modeling quantile processes. This function accepts \code{Gau} to specify 
piecewise Gaussian functions and \code{Gamma} to specify piecewise Gamma functions.}

\item{basis.shape}{a number specifying the shape parameter of the 
piecewise Gamma functions if \code{basis.fun = Gamma}, the default value is 5.}

\item{L}{a number specifying the number of of basis functions used in the 
basis expansion for modeling exposure quantile functions.}

\item{niter}{a number specifying the number of MCMC iterations.}

\item{inde}{a logical value indicating whether exposure quantile processes are 
assumed to be independent across time points. 
\code{inde = TRUE}, this function estimates basis coefficients assuming 
exposure quantile processes are independent across time points; 
\code{inde = FALSE} this function estimates basis coefficients assuming 
exposure quantile processes are temporally correlated by introducing 1-st order
Gaussian Markov random process for the unconstrained latent basis coefficients.}
}
\value{
If \code{inde = TRUE}, this function returns a list contains: (1) a matrix containing samples of 
basis coefficients (\code{parm}), (2) a matrix containing samples of unconstrained latent basis coefficients 
(\code{theta.star}), and (3) a matrix containing whether the MCMC sample was accepted in MH algorithms (\code{accept}), 
and (3) a matrix containing samples of unconstrained latent variables introduced for basis coefficients; 

if \code{inde = FALSE}, this function returns a list contains: (1) a matrix containing 
samples of overall basis coefficients (\code{theta.overall}), (2) a matrix containing 
samples of parameters related to introduce temporal dependency of quantile processes (\code{parm}), 
(3) a list of length \code{L} with each element includes a matrix containing samples of time-specific 
\code{l}-th unconstrained latent basis coefficients (\code{theta.star.time}), 
(4) a matrix containing time-specific basis coefficients (i.e., intercept included in the basis expansion)
(\code{theta0.time}).
}
\description{
This function estimates basis coefficients of basis functions used for expanding 
exposure quantile functions using 
semiparametric Bayesian approaches based on individual-level exposures.
}
\examples{
# Simulate time-specific means and SDs of exposure quantile functions
mean.vec <- rnorm(1000, mean = 7.2, sd = 1)
sd.vec <- rnorm(1000, mean = 1, sd = 0.2)

# Simulate individual-level exposures 
x.sim.mat <- do.call(rbind, lapply(1:1000, 
function(x) rnorm(num.size.time, mean.vec[x], sd.vec[x])))

# Estimate exposure quantile functions at the time point 1 
  re.fit.exp.ind <- fit.exposure(x.ind = x.sim.mat[1, ],
  basis.fun = "Gau", L = 4, niter = 10000, inde = T)

}
\references{
Reich, B. J. (2012). Spatiotemporal quantile regression for detecting 
distributional changes in environmental processes. 
\emph{Journal of the Royal Statistical Society: Series C (Applied Statistics), 61(4)}, 535-553.
}
