% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs_healthmodel.R
\name{fit.health.quan.errors}
\alias{fit.health.quan.errors}
\title{Fit negative-binomial regression using estimated exposure quantile functions as the covariate}
\usage{
fit.health.quan.errors(
  y,
  Z.design = NULL,
  L,
  basis.fun,
  basis.shape,
  n,
  theta.pri.mu,
  theta.pri.pre,
  rand.int = FALSE,
  str = NULL,
  niter = 5000,
  burn_in = 2500
)
}
\arguments{
\item{y}{a vector containing observed aggregate health outcomes.}

\item{Z.design}{a design matrix containing confounders that are controlled for.}

\item{L}{a number specifying the number of of basis functions that has been used in the 
basis expansion for modeling exposure quantile functions.}

\item{basis.fun}{a character string specifying which basis functions that have been used 
for modeling quantile processes. This function accepts \code{Gau} to specify 
piecewise Gaussian functions and \code{Gamma} to specify piecewise Gamma functions.}

\item{basis.shape}{a number specifying the shape parameter of the 
piecewise Gamma functions if \code{basis.fun = Gamma}, the default value is 5.}

\item{n}{a number specifying the degrees of orthonormal Bernstein basis functions 
that are used for expanding \eqn{\beta(\tau)}.}

\item{theta.pri.mu}{a list of length \eqn{T} (i.e., the number of time points) 
with each element including the mean of the multivariate normal (MVN) 
prior assumed for basis coefficients that has been used to model exposure quantile functions.}

\item{theta.pri.pre}{a list with each element including 
the precision matrix of the MVN prior.}

\item{str}{a character string specifying the structure of unit-specific random intercepts 
when \code{rand.int} = \code{TRUE}. \code{AR-1} fits random walk model of order 1 
and \code{exch} specifies exchangeable unit-specific random intercepts.}

\item{niter}{a number specifying the number of MCMC iterations, the default is 5000.}

\item{burn_in}{a number specifying the number of iterations that is discarded, the default is 2500.}
}
\value{
This function returns a list containing: (1) a matrix containing samples of 
basis coefficients of orthonormal Bernstein polynomials and coefficients of controlled confounders, 
(2) WAIC, and (3) parameters related to the mean-zero residual process (if \code{rand.int} = \code{TRUE}).
}
\description{
This function fits negative binomial regression models using estimated 
exposure quantile functions as the covariate while accounting for uncertainties associated 
with estimating exposure quantile functions by introducing MVN prior for basis coefficients 
used for modeling exposure quantile functions to analyze 
aggregate health outcomes in environmental epidemiological studies with a time-series design.
}
\examples{
# Simulate time-specific means and SDs of exposure quantile functions
mean.vec <- rnorm(10, mean = 7.2, sd = 1)
sd.vec <- rnorm(10, mean = 1, sd = 0.2)

# Specify the true beta(tau)
betatau.true <- function(x) {x}

# Generate aggregate health outcomes 
integral <- function(x, betatau.fun, mean, sd){
  re = betatau.fun(x)*qnorm(x, mean, sd)
  return(re)
}

integration.vec <- do.call(rbind,
lapply(1:10, function(y) integrate(integral, lower = 0, upper = 1,
betatau.fun = betatau.true, mean = mean.vec[y], sd = sd.vec[y])$value))

y.vec <- apply(cbind(rep(20, 10), 1/(1+exp(-3.5 + integration.vec))), 1,
function(x) rnbinom(1, size = x[1], prob = x[2]))

# Estimate exposure quantile functions assuming exposure quantile functions 
# are independent across time points.
re.fit.exp <- NULL
for(i in 1:10){
re.fit.exp[[i]] <- fit.exposure(x.ind = x.sim.mat[i, ],
basis.fun = "Gau", L = 4, niter = 10000, inde = T) }

# Obtain means and precision matrices of the MVN prior.
MVNprior.mu.pre.ind <- exposure.prior(re = re.fit.exp, burn_in = 5000, inde = T)

re.fit.quan.errors <- fit.health.quan.errors(y = y.vec, 
L = 4, basis.fun = "Gau", n = 1, 
theta.pri.mu = MVNprior.mu.pre.ind$theta.pri.mu, 
theta.pri.pre = MVNprior.mu.pre.ind$theta.pri.pre,
rand.int = FALSE,
niter = 5000, burn_in = 2500)

}
\references{
Bellucci, M. A. (2014). On the explicit representation of orthonormal Bernstein polynomials. 
\emph{arXiv preprint} \href{https://arxiv.org/abs/1404.2293}{\emph{arXiv:1404.2293}}.
}
