% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs_healthmodel.R
\name{get_beta_bernstein}
\alias{get_beta_bernstein}
\title{Compute basis coefficients of orthonormal Bernstein polynomials used for
approximating a given function}
\usage{
get_beta_bernstein(j, n, f)
}
\arguments{
\item{j}{a number specifying the index of the basis coefficient, \eqn{j=0,\dots,n}.}

\item{n}{a number specifying the degrees of orthonormal Bernstein polynomials.}

\item{f}{the name of self-defined R function that to be approximated.}
}
\description{
This function computes basis coefficients of orthonormal Bernstein polynomials used for
approximating a given function following Eqn. (22) provided in Bellucci (2014). 
Please refer to Section 5 Function Approcimation in Bellucci (2014) for more details.
}
\examples{
# Compute basis coefficients of orthonormal Bernstein polynomials used 
# for approximating the linear function f(x) = x
beta.true <- sapply(0:1, function(x) get_beta_bernstein(x, n = 1, f = function(y){y}))

}
\references{
Bellucci, M. A. (2014). On the explicit representation of orthonormal Bernstein polynomials. 
\emph{arXiv preprint} \href{https://arxiv.org/abs/1404.2293}{\emph{arXiv:1404.2293}}.
}
