% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs_quan.R
\name{quan.fun}
\alias{quan.fun}
\title{Quantile function specified via the basis expansion}
\usage{
quan.fun(tau, L, theta.vec, alpha, basis.fun, shape = 5)
}
\arguments{
\item{tau}{a vector of probabilities (i.e., percentile levels).}

\item{L}{a number specifying the number of of basis functions used in the 
basis expansion for defining the quantile function.}

\item{theta.vec}{a vector of lenght \code{L} containing basis coefficients determing 
the shape of the distribution.}

\item{alpha}{a numeric value specifying the median of the distribution.}

\item{basis.fun}{a character string specifying which basis functions are used 
for modeling quantile processes. This function accepts \code{Gau} to specify 
piecewise Gaussian functions and \code{Gamma} to specify piecewise Gamma functions.}

\item{shape}{a number specifying the shape parameter of the 
piecewise Gamma functions if \code{basis.fun = Gamma}, the default value is 5.}
}
\description{
This function evaluates the quantile function specified via the basis expansion using 
piecewise Gaussian functions or piecewise Gamma functions.
}
\examples{
# Evalute the quantile function of a random variable following a normal distribution with 
# mean 0 and variance 1, but specify its quantile function via the basis expansion using 
# 4 piecewise Gaussian functions. 
quan.fun(tau = seq(0, 1, 0.2), L = 4, theta.vec = rep(1, 4), alpha = 0, basis.fun = "Gau")
# This function should return the same value as applying the function
qnorm(p = seq(0, 1, 0.2), mean = 0, sd = 1)

}
